more code cleanups and warning removals
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@173

+++ 3dc/win95/win_func.cpp
@@ -75,8 +75,8 @@ BOOL g_bMustRedrawScreen = FALSE;
 
 void CheckForWindowsMessages(void)
 {
-	MSG         msg;
-	extern signed int MouseWheelStatus;
+	MSG	msg;
+	extern int16_t MouseWheelStatus;
 	
 	MouseWheelStatus = 0;
